requirements.txt
twitchio==2.9.1
yandex-music==2.1.1
python-dotenv==1.0.0
aiohttp==3.9.1
